pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "discord-autobot"
version = "0.1.0"
dependencies = [
  "discord.py"
]